+++ include/Event.h
@@ -6,13 +6,13 @@ class Event {
 public:
   static unsigned long totalEvents;
 
-  Event(float time, EventType eventType);
+  Event(unsigned long time, EventType eventType);
   unsigned long getId() const;
-  float getTime() const;
+  unsigned long getTime() const;
   EventType getType() const;
 
 protected:
   unsigned long _id;
-  float _eventTime;
+  unsigned long _eventTime;
   EventType _eventType;
 };
