+++ src/ElevatorArrival.cpp
@@ -1,7 +1,7 @@
 #include "ElevatorArrival.h"
 #include "EventType.h"
 
-ElevatorArrival::ElevatorArrival(const float eventTime, const std::shared_ptr<const Elevator> elevator, const std::shared_ptr<const Floor> floor)
+ElevatorArrival::ElevatorArrival(const unsigned long eventTime, const std::shared_ptr<const Elevator> elevator, const std::shared_ptr<const Floor> floor)
  : Event(eventTime, EventType::elevatorArrival)
  , _elevator(elevator)
  , _floor(floor) {}
